Fikser component-preview for parts i layout (setter inn manglende editorView felt). Diverse tweaks på alert-box
