+++ src/components.less
@@ -12,7 +12,7 @@
 @import './components/macros/infoboks/MacroInfoBoks.less';
 @import './components/macros/quote/MacroQuote.less';
 @import './components/macros/varselboks/MacroVarselBoks.less';
-@import './components/macros/Video/MacroVideo.less';
+@import 'src/components/macros/video/MacroVideo.less';
 @import './components/pages/error-page/errorcode-content/clear-icon/ClearIcon.less';
 @import './components/pages/error-page/errorcode-content/ErrorContent404.less';
 @import './components/pages/error-page/errorcode-content/search-form/SearchForm.less';
