+++ src/components/_common/alert-box/AlertBox.module.scss
@@ -10,4 +10,8 @@
     * > p:last-child {
         margin-bottom: 0;
     }
+
+    &:global(.navds-alert--small > .navds-alert__icon) {
+        align-self: center;
+    }
 }
