+++ src/components/macros/infoboks/MacroInfoBoks.tsx
@@ -3,6 +3,7 @@ import { MacroInfoBoksProps } from '../../../types/macro-props/infoBoks';
 import { AlertBox } from '../../_common/alert-box/AlertBox';
 import { ParsedHtml } from '../../_common/parsed-html/ParsedHtml';
 
+// This macro is deprecated
 export const MacroInfoBoks = ({ config }: MacroInfoBoksProps) => {
     if (!config?.infoBoks) {
         return null;
